+++ src/app/core/account-actions-deactivate/account-actions-deactivate.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http'
 import { Injectable } from '@angular/core'
 import { Observable } from 'rxjs'
-import { catchError, retry } from 'rxjs/operators'
+import { catchError } from 'rxjs/operators'
 
 import { ErrorHandlerService } from '../error-handler/error-handler.service'
 import {
@@ -34,10 +34,7 @@ export class AccountActionsDeactivateService {
       .get<ExpiringLinkVerification>(
         runtimeEnvironment.API_WEB + `account/deactivate/` + token
       )
-      .pipe(
-        retry(3),
-        catchError((error) => this._errorHandler.handleError(error))
-      )
+      .pipe(catchError((error) => this._errorHandler.handleError(error)))
   }
 
   deactivateAccount(
@@ -49,10 +46,7 @@ export class AccountActionsDeactivateService {
         runtimeEnvironment.API_WEB + `account/deactivate/` + token,
         data
       )
-      .pipe(
-        retry(3),
-        catchError((error) => this._errorHandler.handleError(error))
-      )
+      .pipe(catchError((error) => this._errorHandler.handleError(error)))
   }
 
   generateDeactivationLink(): Observable<DeactivateResponse> {
@@ -61,9 +55,6 @@ export class AccountActionsDeactivateService {
         runtimeEnvironment.API_WEB + `account/send-deactivate-account.json`,
         this.options
       )
-      .pipe(
-        retry(3),
-        catchError((error) => this._errorHandler.handleError(error))
-      )
+      .pipe(catchError((error) => this._errorHandler.handleError(error)))
   }
 }